Library

Documentation for AstroPropagators.jl.

AstroPropagators.Cowell_EOM! — Method

function Cowell_EOM!(
    du::AbstractVector,
    u::AbstractVector,
    p::ComponentVector,
    t::Number,
    models::AbstractDynamicsModel,
)

Cowell propagation schema for orbital trajectories

Arguments: -du::AbstractVector: In-place vector to store the instantenous rate of change of the current state with respect to time. -u::AbstractVector: The current Cartesian state. -p::ComponentVector: The parameter vector, only the simulation start date JD is provided. -t::Number: The current time. -models::AbstractDynamicsModel: Tuple of the acceleration models.

Returns:

nothing

AstroPropagators.Cowell_EOM — Method

function Cowell_EOM(
    u::AbstractVector,
    p::ComponentVector,
    t::Number,
    models::NTuple{N,AstroForceModels.AbstractAstroForceModel},
) where {N}

Cowell propagation schema for orbital trajectories

Arguments: -u::AbstractVector: The current Cartesian state. -p::ComponentVector: The parameter vector, only the simulation start date JD is provided. -t::Number: The current time. -models::NTuple{N,AstroForceModels.AbstractAstroForceModel}: Tuple of the acceleration models.

Returns: -du::AbstractVector: Instantenous rate of change of the current state with respect to time.

AstroPropagators.EDromo_EOM! — Method

function EDromo_EOM!( du::AbstractVector, u::AbstractArray, p::ComponentVector, ϕ::Number, models::AbstractDynamicsModel, config::RegularizedCoordinateConfig, )

Computes the equations of motion for the EDromo formulation. The formulation is based on [1] and [2].

Arguments

-du::AbstractVector: In-place vector to store the instantenous rate of change of the current state with respect to time. -u::AbstractArray: The EDromo state vector [ζ₁, ζ₂, ζ₃, ζ₄, ζ₅, ζ₆, ζ₇, ζ₈]. -p::ComponentVector: A ComponentVector containing parameters for the propagation, including the gravitational parameter μ and the initial Julian Date JD. -ϕ::Number: The independent variable, which is the fictitious time. -models::AbstractDynamicsModel: Tuple of the acceleration models. -config::RegularizedCoordinateConfig: Configuration struct containing EDromo formulation parameters.

References

[1] Baù, G., Bombardelli, C., Peláez, J., and Lorenzini, E., "Nonsingular orbital elements for special perturbations in the two-body problem". MNRAS 454(3), pp. 2890-2908. 2015. [2] Amato, Davide. "THALASSA: Orbit propagator for near-Earth and cislunar space." Astrophysics Source Code Library (2019): ascl-1905.

AstroPropagators.EDromo_EOM — Method

function EDromo_EOM( u::AbstractArray, p::ComponentVector, ϕ::Number, models::AbstractDynamicsModel, config::RegularizedCoordinateConfig, )

Computes the equations of motion for the EDromo formulation. The formulation is based on [1] and [2].

Arguments

-u::AbstractArray: The EDromo state vector [ζ₁, ζ₂, ζ₃, ζ₄, ζ₅, ζ₆, ζ₇, ζ₈]. -p::ComponentVector: A ComponentVector containing parameters for the propagation, including the gravitational parameter μ and the initial Julian Date JD. -ϕ::Number: The independent variable, which is the fictitious time. -models::AbstractDynamicsModel: Tuple of the acceleration models. -config::RegularizedCoordinateConfig: Configuration struct containing EDromo formulation parameters.

References

[1] Baù, G., Bombardelli, C., Peláez, J., and Lorenzini, E., "Nonsingular orbital elements for special perturbations in the two-body problem". MNRAS 454(3), pp. 2890-2908. 2015. [2] Amato, Davide. "THALASSA: Orbit propagator for near-Earth and cislunar space." Astrophysics Source Code Library (2019): ascl-1905.

AstroPropagators.EDromo_burn — Method

EDromo_burn(burn_time, ΔV, config)

Returns a ContinuousCallback which triggers an impulsive_burn_edromo! maneuver at a specified burn_time.

AstroPropagators.EDromo_time_condition — Method

EDromo_time_condition(u, ϕ, integrator, config; event_time=0.0)

Event function for DifferentialEquations.jl callbacks which triggers when the current integration time equals a specified event_time.

AstroPropagators.GaussVE_EOM! — Method

function GaussVE_EOM!(
    du::AbstractVector,
    u::AbstractVector,
    p::ComponentVector,
    t::Number,
    models::AbstractDynamicsModel,
)

Gauss variational propagation schema for orbital trajectories

Arguments: -du::AbstractVector: In-place vector to store the instantenous rate of change of the current state with respect to time. -u::AbstractVector: The current Keplerian state. -p::ComponentVector: The parameter vector, the simulation start date JD and the central body gravitational parameter. -t::Number: The current time. -models::NTuple{N,AstroForceModels.AbstractAstroForceModel}: Tuple of the acceleration models.

Returns:

nothing

AstroPropagators.GaussVE_EOM — Method

function GaussVE_EOM(
    u::AbstractVector,
    p::ComponentVector,
    t::Number,
    models::AbstractDynamicsModel,
)

Gauss variational propagation schema for orbital trajectories

Arguments: -u::AbstractVector: The current Keplerian state. -p::ComponentVector: The parameter vector, the simulation start date JD and the central body gravitational parameter. -t::Number: The current time. -models::NTuple{N,AstroForceModels.AbstractAstroForceModel}: Tuple of the acceleration models.

Returns: -du::AbstractVector: Instantenous rate of change of the current state with respect to time.

AstroPropagators.KS_burn — Method

KS_burn(burn_time, ΔV, config)

Returns a ContinuousCallback which triggers an impulsive_burn_ks! maneuver at a specified burn_time.

AstroPropagators.KS_time_condition — Method

KS_time_condition(u, ϕ, integrator, config; event_time=0.0)

Event function for DifferentialEquations.jl callbacks which triggers when the current integration time equals a specified event_time.

AstroPropagators.Milankovich_EOM! — Method

function Milankovich_EOM!(
    du::AbstractVector,
    u::AbstractVector,
    p::ComponentVector,
    t::Number,
    models::AbstractDynamicsModel,
)

Milankovich propagation schema for orbital trajectories

Arguments: -du::AbstractVector: In-place vector to store the instantenous rate of change of the current state with respect to time. -u::AbstractVector: The current Milankovich state. -p::ComponentVector: The parameter vector, the simulation start date JD and the central body gravitational parameter. -t::Number: The current time. -models::AbstractDynamicsModel: Tuple of the acceleration models.

Returns:

nothing

AstroPropagators.Milankovich_EOM — Method

function Milankovich_EOM(
    u::AbstractVector,
    p::ComponentVector,
    t::Number,
    models::AbstractDynamicsModel,
)

Milankovich propagation schema for orbital trajectories

Arguments: -u::AbstractVector: The current Milankovich state. -p::ComponentVector: The parameter vector, the simulation start date JD and the central body gravitational parameter. -t::Number: The current time. -models::AbstractDynamicsModel: Tuple of the acceleration models.

Returns: -du::AbstractVector: Instantenous rate of change of the current state with respect to time.

AstroPropagators.StiSche_burn — Method

StiSche_burn(burn_time, ΔV, config)

Returns a ContinuousCallback which triggers an impulsive_burn_stische! maneuver at a specified burn_time.

AstroPropagators.StiSche_time_condition — Method

StiSche_time_condition(u, ϕ, integrator, config; event_time=0.0)

Event function for DifferentialEquations.jl callbacks which triggers when the current integration time equals a specified event_time.

AstroPropagators.USM6_EOM! — Method

function USM6_EOM!(
    du::AbstractVector,
    u::AbstractVector,
    p::ComponentVector,
    t::Number,
    models::AbstractDynamicsModel,
)

Unified State Model (MRP's) propagation schema for orbital trajectories

Arguments: -du::AbstractVector: In-place vector to store the instantenous rate of change of the current state with respect to time. -u::AbstractVector: The current USM6 state. -p::ComponentVector: The parameter vector, the simulation start date JD and the central body gravitational parameter. -t::Number: The current time. -models::AbstractDynamicsModel: Tuple of the acceleration models.

Returns:

nothing

AstroPropagators.USM6_EOM — Method

function USM6_EOM(
    u::AbstractVector,
    p::ComponentVector,
    t::Number,
    models::AbstractDynamicsModel,
)

Unified State Model (MRP's) propagation schema for orbital trajectories

Arguments: -u::AbstractVector: The current USM6 state. -p::ComponentVector: The parameter vector, the simulation start date JD and the central body gravitational parameter. -t::Number: The current time. -models::AbstractDynamicsModel: Tuple of the acceleration models.

Returns: -du::AbstractVector: Instantenous rate of change of the current state with respect to time.

AstroPropagators.USM7_EOM! — Method

function USM7_EOM!(
    du::AbstractVector,
    u::AbstractVector,
    p::ComponentVector,
    t::Number,
    models::AbstractDynamicsModel,
)

Unified State Model (quaternions) propagation schema for orbital trajectories

Arguments: -du::AbstractVector: In-place vector to store the instantenous rate of change of the current state with respect to time. -u::AbstractVector: The current USM7 state. -p::ComponentVector: The parameter vector, the simulation start date JD and the central body gravitational parameter. -t::Number: The current time. -models::AbstractDynamicsModel: Tuple of the acceleration models.

Returns:

nothing

AstroPropagators.USM7_EOM — Method

function USM7_EOM(
    u::AbstractVector,
    p::ComponentVector,
    t::Number,
    models::AbstractDynamicsModel,
)

Unified State Model (quaternions) propagation schema for orbital trajectories

Arguments: -u::AbstractVector: The current USM7 state. -p::ComponentVector: The parameter vector, the simulation start date JD and the central body gravitational parameter. -t::Number: The current time. -models::AbstractDynamicsModel: Tuple of the acceleration models.

Returns: -du::AbstractVector: Instantenous rate of change of the current state with respect to time.

AstroPropagators.USMEM_EOM! — Method

function USMEM_EOM!(
    du::AbstractVector,
    u::AbstractVector,
    p::ComponentVector,
    t::Number,
    models::AbstractDynamicsModel,
)

Unified State Model (exponential mapping) propagation schema for orbital trajectories

Arguments: -du::AbstractVector: In-place vector to store the instantenous rate of change of the current state with respect to time. -u::AbstractVector: The current USM7 state. -p::ComponentVector: The parameter vector, the simulation start date JD and the central body gravitational parameter. -t::Number: The current time. -models::AbstractDynamicsModel: Tuple of the acceleration models.

Keyword Arguments"

-Φ_tol::Float64: The value to switch to the Taylor series expansion to avoid singularity.

Returns:

nothing

AstroPropagators.USMEM_EOM — Method

function USMEM_EOM(
    u::AbstractVector,
    p::ComponentVector,
    t::Number,
    models::AbstractDynamicsModel,
)

Unified State Model (exponential mapping) propagation schema for orbital trajectories

Arguments: -u::AbstractVector: The current USMEM state. -p::ComponentVector: The parameter vector, the simulation start date JD and the central body gravitational parameter. -t::Number: The current time. -models::AbstractDynamicsModel: Tuple of the acceleration models.

Keyword Arguments"

-Φ_tol::Float64: The value to switch to the Taylor series expansion to avoid singularity.

Returns: -du::AbstractVector: Instantenous rate of change of the current state with respect to time.

AstroPropagators.end_EDromo_integration — Method

end_EDromo_integration(stop_time, config)

Returns a ContinuousCallback which terminates a DifferentialEquations.jl integration when the current time equals event_time.

AstroPropagators.end_KS_integration — Method

end_KS_integration(stop_time, config)

Returns a ContinuousCallback which terminates a DifferentialEquations.jl integration when the current time equals event_time.

AstroPropagators.end_StiSche_integration — Method

end_StiSche_integration(stop_time, config)

Returns a ContinuousCallback which terminates a DifferentialEquations.jl integration when the current time equals event_time.

AstroPropagators.impulsive_burn! — Method

function impulsive_burn!(
    integrator::SciMLBase.DEIntegrator, 
    ΔV::AbstractVector; 
    coordinate_set::AstroCoords.AstroCoord=Cartesian
)

Computes new state from an input impulsive burn. The supplied burn should be in the inertial frame and has to be computed before-hand. Both will change in future iterations.

Arguments: -integrator::SciMLBase.DEIntegrator: The differential equation integrator object. -ΔV::AbstractVector: The deltaV of the impulsive burn. -coordinate_set::AstroCoords.AstroCoord=Cartesian: The coordinate set the propagation is occurring in. Returns: -nothing

AstroPropagators.impulsive_burn_edromo! — Method

impulsive_burn_edromo!(integrator, ΔV, config)

Applies an impulsive maneuver ΔV to an EDromo state within a DifferentialEquations.jl integrator.

AstroPropagators.impulsive_burn_ks! — Method

impulsive_burn_ks!(integrator, ΔV, config)

Applies an impulsive maneuver ΔV to a KS state within a DifferentialEquations.jl integrator.

AstroPropagators.impulsive_burn_stische! — Method

impulsive_burn_stische!(integrator, ΔV, config)

Applies an impulsive maneuver ΔV to a Stiefel-Scheifele state within a DifferentialEquations.jl integrator.

AstroPropagators.inert2orb_edromo — Method

inert2orb_edromo(v_inertial, u, cν, sν) -> SVector{3}

Transforms a vector from the inertial frame to the EDROMO orbital frame.

The orbital frame is defined by the radial, tangential, and normal directions.

Args

v_inertial::AbstractVector: The vector in the inertial reference frame.

u::AbstractArray: The EDromo state vector [ζ₁, ζ₂, ζ₃, ζ₄, ζ₅, ζ₆, ζ₇, ζ₈].

cν::Number: The cosine of the auxiliary angle ν.

sν::Number: The sine of the auxiliary angle ν.

Returns

SVector{3}: The transformed vector in the EDROMO orbital frame.